Throw out the binary. Why was this in the repo in the first place?

Conflicts:
	tdfc

+++ src/ccdfgbody.cc
@@ -855,7 +855,7 @@ void ccdfgprocrun(ofstream *fout, string name, Operator *op,
 			    << loc
 			    << "=STREAM_EOS(" 
 			    << "in[" 
-			    << (int)(ispec->getStream()->getAnnote(CC_STREAM_ID)) 
+			    << (long)(ispec->getStream()->getAnnote(CC_STREAM_ID)) 
 			    << "]" 
 			    << ");" << endl;
 
@@ -864,7 +864,7 @@ void ccdfgprocrun(ofstream *fout, string name, Operator *op,
 			    << loc
 			    << "=STREAM_EOFR("
 			    << "in["
-			    << (int)(ispec->getStream()->getAnnote(CC_STREAM_ID))
+			    << (long)(ispec->getStream()->getAnnote(CC_STREAM_ID))
 			    << "]"
 			    << ");" << endl;
 
@@ -904,7 +904,7 @@ void ccdfgprocrun(ofstream *fout, string name, Operator *op,
 		      *fout << "          " 
 			    << ispec->getStream()->getName() 
 			    << (floattyp? "=STREAM_READ_FLOAT(" : (doubletyp)? "=STREAM_READ_DOUBLE(" : "=STREAM_READ_NOACC(") 
-			    << "in[" << (int)(ispec->getStream()->getAnnote(CC_STREAM_ID)) 
+			    << "in[" << (long)(ispec->getStream()->getAnnote(CC_STREAM_ID)) 
 			    << "]"
 			    << ");" << endl;
 		    }
@@ -912,17 +912,17 @@ void ccdfgprocrun(ofstream *fout, string name, Operator *op,
 		    {
 		    *fout << "          "
 			  << "if (!input_free[" 
-			  << (int)(ispec->getStream()->getAnnote(CC_STREAM_ID)) 
+			  << (long)(ispec->getStream()->getAnnote(CC_STREAM_ID)) 
 			  << "])" << endl;
 		      *fout << "          " 
 			    << "STREAM_FREE(" 
-			    << "in[" << (int)(ispec->getStream()->getAnnote(CC_STREAM_ID)) 
+			    << "in[" << (long)(ispec->getStream()->getAnnote(CC_STREAM_ID)) 
 			    << "]"
 			    << ");" << endl;
-		       early_free[(int)(ispec->getStream()->getAnnote(CC_STREAM_ID))]=1;
+		       early_free[(long)(ispec->getStream()->getAnnote(CC_STREAM_ID))]=1;
 		       *fout << "          " 
 			     << "input_free[" 
-			     << (int)(ispec->getStream()->getAnnote(CC_STREAM_ID))
+			     << (long)(ispec->getStream()->getAnnote(CC_STREAM_ID))
 			     << "]=1;" << endl;
 		    }
 			  // Special case of End-of-Frame
@@ -937,7 +937,7 @@ void ccdfgprocrun(ofstream *fout, string name, Operator *op,
 					// dummy read of the token..
 				    *fout << "          "
 					    << (floattyp? "STREAM_READ_FLOAT(" : (doubletyp)? "STREAM_READ_DOUBLE(" : "STREAM_READ_NOACC(")
-					    << "in[" << (int)(ispec->getStream()->getAnnote(CC_STREAM_ID))
+					    << "in[" << (long)(ispec->getStream()->getAnnote(CC_STREAM_ID))
 					    << "]"
 					    << ");" << endl;
 
@@ -1119,7 +1119,7 @@ void ccdfgprocrun(ofstream *fout, string name, Operator *op,
 							bool floattype = (outType==TYPE_FLOAT);
 						        bool doubletype = (outType==TYPE_DOUBLE);
 
-							int id=(int)(ssym->getAnnote(CC_STREAM_ID));
+							int id=(long)(ssym->getAnnote(CC_STREAM_ID));
 							*fout << "          " ;
 							*fout  << (floattype?"STREAM_WRITE_FLOAT":doubletype?"STREAM_WRITE_DOUBLE":"STREAM_WRITE_NOACC")
 								<< "(out[" << id << "]," << nodetofout(dfg,n,nodenums) << ");" << endl; // asym->toString();
@@ -1167,19 +1167,19 @@ void ccdfgprocrun(ofstream *fout, string name, Operator *op,
       *fout << "  }" << endl;
       // any final stuff
       if (!noReturnValue(rsym))
-	if (early_close[(int)(rsym->getAnnote(CC_STREAM_ID))])
+	if (early_close[(long)(rsym->getAnnote(CC_STREAM_ID))])
 	  {
 	    *fout <<"  if (!output_close[" 
-		  <<  (int)(rsym->getAnnote(CC_STREAM_ID))
+		  <<  (long)(rsym->getAnnote(CC_STREAM_ID))
 		  << "])" << endl;
 	    *fout << "  STREAM_CLOSE(" 
-		  << "out[" << (int)(rsym->getAnnote(CC_STREAM_ID))
+		  << "out[" << (long)(rsym->getAnnote(CC_STREAM_ID))
 		  << "]" 
 		  << ");" << endl;
 	  }
 	else
 	  *fout << "  STREAM_CLOSE(" 
-		<< "out[" << (int)(rsym->getAnnote(CC_STREAM_ID))
+		<< "out[" << (long)(rsym->getAnnote(CC_STREAM_ID))
 		<< "]"
 		<< ");" << endl;
       forall(sym,*argtypes)
@@ -1189,40 +1189,40 @@ void ccdfgprocrun(ofstream *fout, string name, Operator *op,
 	      SymbolStream *ssym=(SymbolStream *)sym;
 	      if (ssym->getDir()==STREAM_OUT)
 		{
-		  if (early_close[(int)(ssym->getAnnote(CC_STREAM_ID))])
+		  if (early_close[(long)(ssym->getAnnote(CC_STREAM_ID))])
 		    {
 		      *fout <<"  if (!output_close[" 
-			    <<  (int)(ssym->getAnnote(CC_STREAM_ID))
+			    <<  (long)(ssym->getAnnote(CC_STREAM_ID))
 			    << "])" << endl;
 		      *fout << "  STREAM_CLOSE(" 
-			    << "out[" << (int)(ssym->getAnnote(CC_STREAM_ID))
+			    << "out[" << (long)(ssym->getAnnote(CC_STREAM_ID))
 			    << "]" 
 			    << ");" << endl;
 		    }
 		  else
 		    *fout << "  STREAM_CLOSE(" 
-			  << "out[" << (int)(ssym->getAnnote(CC_STREAM_ID)) 
+			  << "out[" << (long)(ssym->getAnnote(CC_STREAM_ID)) 
 			  << "]" 
 			  << ");" << endl;
 		}
 	      else
-		if (early_free[(int)(ssym->getAnnote(CC_STREAM_ID))])
+		if (early_free[(long)(ssym->getAnnote(CC_STREAM_ID))])
 		  {
-		    cerr << "DEBUG early free[" << (int)(ssym->getAnnote(CC_STREAM_ID))
-			 << "]=" << early_free[(int)(ssym->getAnnote(CC_STREAM_ID))]
+		    cerr << "DEBUG early free[" << (long)(ssym->getAnnote(CC_STREAM_ID))
+			 << "]=" << early_free[(long)(ssym->getAnnote(CC_STREAM_ID))]
 			 << endl;
 		    
 		    *fout <<"  if (!input_free[" 
-			  << (int)(ssym->getAnnote(CC_STREAM_ID)) 
+			  << (long)(ssym->getAnnote(CC_STREAM_ID)) 
 			  << "])" << endl;
 		    *fout << "    STREAM_FREE(" 
-			  << "in[" << (int)(ssym->getAnnote(CC_STREAM_ID))
+			  << "in[" << (long)(ssym->getAnnote(CC_STREAM_ID))
 			  << "]" 
 			  << ");" << endl;
 		  }
 		else
 		  *fout << "  STREAM_FREE(" 
-			<< "in[" << (int)(ssym->getAnnote(CC_STREAM_ID))
+			<< "in[" << (long)(ssym->getAnnote(CC_STREAM_ID))
 			<< "]" 
 			<< ");" << endl;
 	    }
